修改了api.js,修正了一个错误：该错误曾导致在登陆后依然会看到所有用户的便利贴，现在，用户登录后只会看到自己的便利贴了

+++ routes/api.js
@@ -1,22 +1,17 @@
 var express = require('express');
 var router = express.Router();
-
 var Note = require('../models/note')
-
-/*GET api listing. */
 
 router.get('/notes', function(req, res, next) {
     var opts = { raw: true }
     if (req.session && req.session.user) {
         opts.where = { uid: req.session.user.id }
     }
-    Note.findAll({ raw: true }).then(notes => {
-        // console.log(notes)
-        res.send({
-            status: 0,
-            data: notes
-        })
-    }).catch(() => {
+
+    Note.findAll(opts).then(function(notes) {
+        console.log(notes)
+        res.send({ status: 0, data: notes });
+    }).catch(function() {
         res.send({ status: 1, errorMsg: 'Database Exceptions' });
     });
 });
@@ -32,7 +27,7 @@ router.post('/notes/add', function(req, res, next) {
 
     var note = req.body.note;
     var uid = req.session.user.id;
-    // console.log({ text: note, uid: uid })
+    console.log({ text: note, uid: uid })
     var username = req.session.user.username;
     var update = new Date().getTime();
 
@@ -42,9 +37,9 @@ router.post('/notes/add', function(req, res, next) {
         username: username,
         createdAt: update,
         updatedAt: update
-    }).then(data => {
+    }).then(function(data) {
         res.send({ status: 0, result: data.get({ plain: true }) })
-    }).catch(() => {
+    }).catch(function() {
         res.send({
             status: 1,
             errorMsg: 'Database Exceptions or You have no permissions'
@@ -65,17 +60,15 @@ router.post('/notes/edit', function(req, res, next) {
     var note = req.body.note;
     var uid = req.session.user.id;
     var update = new Date().getTime();
-
     Note.update({
         text: note,
         updatedAt: update
-    }, { where: { id: noteId, uid: uid }, returning: true, plain: true }).then((lists) => {
+    }, { where: { id: noteId, uid: uid }, returning: true, plain: true }).then(function(lists) {
         if (lists[1] === 0) {
             return res.send({ status: 1, errorMsg: 'You have no permissions' })
         }
 
         res.send({ status: 0 })
-            // console.log('编辑回调', lists)
     }).catch(function(e) {
         res.send({ status: 1, errorMsg: 'Database Exceptions or You have no permissions' })
     })
@@ -90,9 +83,7 @@ router.post('/notes/delete', function(req, res, next) {
     var noteId = req.body.id;
     var uid = req.session.user.id;
 
-    Note.destroy({
-        where: { id: noteId, uid: uid }
-    }).then(deleteLen => {
+    Note.destroy({ where: { id: noteId, uid: uid } }).then(function(deleteLen) {
         if (deleteLen === 0) {
             return res.send({ status: 1, errorMsg: 'You have no permissions' })
         }
